Add files via upload
Pulled from my other GitHub. It is code in Mathematica that visualizes Riemann sums for 3D areas of revolution. The user inputs the function, the domain, as well as DeltaX (the spacing between each "cut" of the integral.)

The tool was meant to help students understand the visuals of 3D bodies of revolution whilst also connecting the idea of "infinitely small" divisions and integrals.
This project was made as part of an honors project for my Calculus II class.